Remove fdvar_forceNeqInline
=============================== Coverage summary ================
Statements   : 97.4% ( 3071/3153 ), 75 ignored
Branches     : 93.49% ( 1437/1537 ), 76 ignored
Functions    : 98.98% ( 291/294 ), 6 ignored
Lines        : 97.55% ( 2508/2571 )
=================================================================
