+++ src/fdvar.js
@@ -1,26 +1,18 @@
 import {
-  EMPTY,
   NO_CHANGES,
-  REJECTED,
   SOME_CHANGES,
 
   ASSERT,
   ASSERT_DOMAIN,
   ASSERT_UNUSED_DOMAIN,
-  ASSERT_DOMAIN_EMPTY_CHECK,
 } from './helpers';
 
 import {
-  domain_getValue,
-  domain_isRejected,
-  domain_isSolved,
   domain_numarr,
   domain_removeGteInline,
   domain_removeGteNumbered,
   domain_removeLteInline,
   domain_removeLteNumbered,
-  domain_removeValueInline,
-  domain_removeValueNumbered,
 } from './domain';
 
 // BODY_START
@@ -80,65 +72,11 @@ function fdvar_removeLteInline(fdvar, value) {
   return NO_CHANGES;
 }
 
-function fdvar_forceNeqInline(fdvar1, fdvar2) {
-  let r = NO_CHANGES;
-  let dom1 = fdvar1.dom;
-  let dom2 = fdvar2.dom;
-
-  ASSERT_DOMAIN_EMPTY_CHECK(dom1);
-  ASSERT_DOMAIN_EMPTY_CHECK(dom2);
-
-  if (domain_isSolved(dom1)) {
-    let value = domain_getValue(dom1);
-    if (typeof dom2 === 'number') {
-      let result = domain_removeValueNumbered(dom2, value);
-      if (result === EMPTY) {
-        fdvar2.dom = result;
-        r = REJECTED;
-      } else if (result !== dom2) {
-        fdvar2.dom = result;
-        r = SOME_CHANGES;
-      } else {
-        r = NO_CHANGES;
-      }
-    } else {
-      r = domain_removeValueInline(dom2, value);
-      if (r !== NO_CHANGES) {
-        fdvar2.dom = domain_numarr(dom2);
-      }
-    }
-  } else if (domain_isSolved(dom2)) {
-    let value = domain_getValue(dom2);
-    if (typeof dom1 === 'number') {
-      let result = domain_removeValueNumbered(dom1, value);
-      if (result === EMPTY) {
-        fdvar1.dom = result;
-        r = REJECTED;
-      } else if (result !== dom1) {
-        fdvar1.dom = result;
-        r = SOME_CHANGES;
-      } else {
-        r = NO_CHANGES;
-      }
-    } else {
-      r = domain_removeValueInline(dom1, value);
-      if (r !== NO_CHANGES) {
-        fdvar1.dom = domain_numarr(dom1);
-      }
-    }
-  }
-
-  ASSERT(r === REJECTED || r === NO_CHANGES || r === SOME_CHANGES, 'turning stuff into enum, must be sure about values');
-  ASSERT((r === REJECTED) === (domain_isRejected(fdvar1.dom) || domain_isRejected(fdvar2.dom)), 'if either domain is rejected, r should reflect this already');
-
-  return r;
-}
 
 // BODY_STOP
 
 export {
   fdvar_create,
-  fdvar_forceNeqInline,
   fdvar_removeGteInline,
   fdvar_removeLteInline,
 };
