+++ src/propagators/neq.js
@@ -1,12 +1,21 @@
 import {
+  EMPTY,
+  NO_CHANGES,
+  REJECTED,
+  SOME_CHANGES,
+
+  ASSERT,
+  ASSERT_DOMAIN_EMPTY_CHECK,
+} from '../helpers';
+import {
   domain_getValue,
+  domain_isRejected,
   domain_isSolved,
+  domain_numarr,
+  domain_removeValueInline,
+  domain_removeValueNumbered,
   domain_sharesNoElements,
 } from '../domain';
-
-import {
-  fdvar_forceNeqInline,
-} from '../fdvar';
 
 // BODY_START
 
@@ -16,7 +25,57 @@ import {
  * @returns {*}
  */
 function propagator_neqStepBare(fdvar1, fdvar2) {
-  return fdvar_forceNeqInline(fdvar1, fdvar2);
+  let result = NO_CHANGES;
+  let dom1 = fdvar1.dom;
+  let dom2 = fdvar2.dom;
+
+  ASSERT_DOMAIN_EMPTY_CHECK(dom1);
+  ASSERT_DOMAIN_EMPTY_CHECK(dom2);
+
+  if (domain_isSolved(dom1)) {
+    let value = domain_getValue(dom1);
+    if (typeof dom2 === 'number') {
+      result = domain_removeValueNumbered(dom2, value);
+      if (result === EMPTY) {
+        fdvar2.dom = result;
+        result = REJECTED;
+      } else if (result !== dom2) {
+        fdvar2.dom = result;
+        result = SOME_CHANGES;
+      } else {
+        result = NO_CHANGES;
+      }
+    } else {
+      result = domain_removeValueInline(dom2, value);
+      if (result !== NO_CHANGES) {
+        fdvar2.dom = domain_numarr(dom2);
+      }
+    }
+  } else if (domain_isSolved(dom2)) {
+    let value = domain_getValue(dom2);
+    if (typeof dom1 === 'number') {
+      result = domain_removeValueNumbered(dom1, value);
+      if (result === EMPTY) {
+        fdvar1.dom = result;
+        result = REJECTED;
+      } else if (result !== dom1) {
+        fdvar1.dom = result;
+        result = SOME_CHANGES;
+      } else {
+        result = NO_CHANGES;
+      }
+    } else {
+      result = domain_removeValueInline(dom1, value);
+      if (result !== NO_CHANGES) {
+        fdvar1.dom = domain_numarr(dom1);
+      }
+    }
+  }
+
+  ASSERT(result === REJECTED || result === NO_CHANGES || result === SOME_CHANGES, 'turning stuff into enum, must be sure about values');
+  ASSERT((result === REJECTED) === (domain_isRejected(fdvar1.dom) || domain_isRejected(fdvar2.dom)), 'if either domain is rejected, r should reflect this already');
+
+  return result;
 }
 
 /**
